requirements.txt
# Autogen Core Framework
autogen-agentchat
autogen-core

# Autogen Extensions (MCP e Ollama)
autogen-ext[mcp,ollama]

# Utils
python-dotenv
